API - Clear exception on delete

+++ api/src/main/java/com/gberard/tournament/domain/service/PhaseServiceImpl.java
@@ -1,7 +1,9 @@
 package com.gberard.tournament.domain.service;
 
+import com.gberard.tournament.domain.exception.PhaseInUseException;
 import com.gberard.tournament.domain.model.Phase;
 import com.gberard.tournament.domain.port.input.PhaseService;
+import com.gberard.tournament.domain.port.output.GameRepository;
 import com.gberard.tournament.domain.port.output.PhaseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,9 @@ public class PhaseServiceImpl implements PhaseService {
     @Autowired
     public PhaseRepository phaseRepository;
 
+    @Autowired
+    public GameRepository gameRepository;
+
     @Override
     public Phase create(Phase phase) {
         return phaseRepository.save(phase);
@@ -27,6 +32,10 @@ public class PhaseServiceImpl implements PhaseService {
 
     @Override
     public boolean delete(Phase phase) {
+        if (gameRepository.existsByPhaseId(phase.id())) {
+            throw new PhaseInUseException(phase.id());
+        }
+
         phaseRepository.deleteById(phase.id());
         return true;
     }
